Move `set_verbosity_and_logfile()` to `Messages`
Per suggestion from @kbwestfall, move the  `set_verbosity_and_logfile()` function
into the `Messages` class from the script utilities module.

	modified:   pypeit/pypmsgs.py
	modified:   pypeit/scripts/coadd_1dspec.py
	modified:   pypeit/scripts/coadd_2dspec.py
	modified:   pypeit/scripts/coadd_datacube.py
	modified:   pypeit/scripts/collate_1d.py
	modified:   pypeit/scripts/flux_calib.py
	modified:   pypeit/scripts/sensfunc.py
	modified:   pypeit/scripts/skysub_regions.py
	modified:   pypeit/scripts/tellfit.py
	modified:   pypeit/scripts/trace_edges.py
	modified:   pypeit/scripts/utils.py
