+++ pypeit/pypmsgs.py
@@ -5,6 +5,7 @@ Module for terminal and file logging.
     Why not use pythons native logging package?
 
 """
+import datetime
 import sys
 import os
 import getpass
@@ -310,5 +311,29 @@ class Messages:
         self._black_YL = ''
         self._yellow_BK = ''
 
+    def set_verbosity_and_logfile(self, scriptname, verbosity):
+        """
+        Set the verbosity level and (optionally) logfile for a script run.
 
+        PypeIt scripts (with the exception of run_pypeit) default to verbosity
+        level = 1.  For certain scripts, having a more verbose output (with an
+        accompanying log file) would be helpful for debugging purposes.  This
+        function provides the ability to set the ``msgs`` verbosity and create
+        a log file for those certain scripts.
 
+        Log filenames have the form scriptname_YYYYMMDD_HHMM.log to differentiate
+        between different runs of the script.  Timestamp is UT.
+
+        Args:
+            scriptname (:obj:`str`, optional):
+                The name of the calling script for use in the logfile
+            verbosity (:obj:`int`, optional):
+                The requested verbosity, passed in from the argument parser.
+                Verbosity level between 0 [none] and 2 [all]
+        """
+        # Create a UT timestamp (to the minute) for the log filename
+        timestamp = datetime.datetime.utcnow().strftime("%Y%m%d-%H%M")
+        # Create a logfile only if verbosity == 2
+        logname = f"{scriptname}_{timestamp}.log" if verbosity == 2 else None
+        # Set the verbosity in msgs
+        self.reset(log=logname, verbosity=verbosity)
