+++ pypeit/scripts/skysub_regions.py
@@ -38,6 +38,7 @@ class SkySubRegions(scriptbase.ScriptBase):
 
         import os
 
+        from pypeit import msgs
         from pypeit.core.gui.skysub_regions import SkySubGUI
         from pypeit.core import flexure
         from pypeit.scripts import utils
@@ -45,7 +46,7 @@ class SkySubRegions(scriptbase.ScriptBase):
         from pypeit.images import buildimage
 
         # Set the verbosity, and create a logfile if verbosity == 2
-        utils.set_verbosity_and_logfile('skysub_regions', args.verbosity)
+        msgs.set_verbosity_and_logfile('skysub_regions', args.verbosity)
 
         # Generate a utilities class
         info = utils.Utilities(None, pypeit_file=args.file, det=args.det)
