+++ pypeit/scripts/tellfit.py
@@ -90,10 +90,9 @@ class TellFit(scriptbase.ScriptBase):
         from pypeit.spectrographs.util import load_spectrograph
         from pypeit.core import telluric
         from pypeit import inputfiles
-        from pypeit.scripts.utils import set_verbosity_and_logfile
 
         # Set the verbosity, and create a logfile if verbosity == 2
-        set_verbosity_and_logfile('tellfit', args.verbosity)
+        msgs.set_verbosity_and_logfile('tellfit', args.verbosity)
 
         # Determine the spectrograph
         header = fits.getheader(args.spec1dfile)
